+++ app/src/main/java/com/zhmu100/ma/ui/components/Greeting.kt
@@ -1,0 +1,13 @@
+package com.zhmu100.ma.ui.components
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun Greeting(name: String, modifier: Modifier = Modifier) {
+    Text(
+        text = "Hello $name!",
+        modifier = modifier
+    )
+}
