+++ app/src/main/java/com/zhmu100/ma/ui/theme/Color.kt
@@ -2,10 +2,8 @@ package com.zhmu100.ma.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
+val DarkGreen = Color(0xFF225453)
+val LightGreen = Color(0xFF4FA997)
+val White = Color(0xFFFFFFFF)
+val LightBlue = Color(0xFF8EA3FF)
+val Orange = Color(0xFFBD5D34)
